# Conflicts:
#	src/AdminConsole/Program.cs

+++ src/AdminConsole/Services/ApplicationService.cs
@@ -2,6 +2,8 @@ using Microsoft.EntityFrameworkCore;
 using Passwordless.AdminConsole.Db;
 using Passwordless.AdminConsole.Models;
 using Passwordless.AdminConsole.Models.DTOs;
+using Passwordless.AdminConsole.Services.MagicLinks;
+using Passwordless.AdminConsole.Services.Mail;
 using Passwordless.AdminConsole.Services.PasswordlessManagement;
 
 namespace Passwordless.AdminConsole.Services;
@@ -10,11 +12,22 @@ public class ApplicationService<TDbContext> : IApplicationService where TDbConte
 {
     private readonly IPasswordlessManagementClient _client;
     private readonly IDbContextFactory<TDbContext> _dbContextFactory;
+    private readonly ISharedBillingService _billingService;
+    private readonly IMailService _mailService;
+    private readonly IMagicLinkBuilder _magicLinkBuilder;
 
-    public ApplicationService(IPasswordlessManagementClient client, IDbContextFactory<TDbContext> dbContextFactory)
+    public ApplicationService(
+        IPasswordlessManagementClient client,
+        IDbContextFactory<TDbContext> dbContextFactory,
+        IMailService mailService,
+        IMagicLinkBuilder magicLinkBuilder,
+        ISharedBillingService billingService)
     {
         _client = client;
         _dbContextFactory = dbContextFactory;
+        _billingService = billingService;
+        _mailService = mailService;
+        _magicLinkBuilder = magicLinkBuilder;
     }
 
     public async Task<MarkDeleteApplicationResponse> MarkApplicationForDeletionAsync(string applicationId, string userName)
@@ -28,14 +41,16 @@ public class ApplicationService<TDbContext> : IApplicationService where TDbConte
 
         if (response.IsDeleted)
         {
-            db.Applications.Entry(application).State = EntityState.Deleted;
+            await DeleteAsync(applicationId);
+            await _mailService.SendApplicationDeletedAsync(application, response.DeleteAt, userName, response.AdminEmails);
         }
         else
         {
             application.DeleteAt = response.DeleteAt;
+            await db.SaveChangesAsync();
+            var magicLink = await _magicLinkBuilder.GetLinkAsync(userName, "/organization/overview");
+            await _mailService.SendApplicationToBeDeletedAsync(application, userName, magicLink, response.AdminEmails);
         }
-
-        await db.SaveChangesAsync();
 
         return response;
     }
@@ -64,9 +79,14 @@ public class ApplicationService<TDbContext> : IApplicationService where TDbConte
     public async Task DeleteAsync(string applicationId)
     {
         await using var db = await _dbContextFactory.CreateDbContextAsync();
-        var application = new Application { Id = applicationId };
+        var application = await db.Applications.SingleOrDefaultAsync(x => x.Id == applicationId);
+        if (application == null) return;
         db.Applications.Remove(application);
-        await db.SaveChangesAsync();
+        var rows = await db.SaveChangesAsync();
+        if (rows > 0 && application.BillingSubscriptionItemId != null)
+        {
+            await _billingService.OnPostApplicationDeletedAsync(application.BillingSubscriptionItemId);
+        }
     }
 
     public async Task CreateApplicationAsync(Application application)
